[Refactor] ContentsTop에서 HubContentTop분리 후 적용
